backend/utils.py
# File: backend/utils.py
def preprocess_image(image):
    from PIL import Image
    import torchvision.transforms as transforms
    transform = transforms.Compose([
        transforms.Resize((224, 224)),
        transforms.ToTensor()
    ])
    return transform(image).unsqueeze(0)