backend/requirements.txt
Flask==3.0.0
torch==2.6.0
torchvision==0.21.0
Pillow==10.0.1
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
matplotlib==3.8.0
opencv-python==4.8.1.78
tensorflow==2.14.0
